The LINE Message API has been moved to features/store and the API route name has been changed.
