--- src/routes/message.route.ts
@@ -1,25 +1,0 @@
-import express from "express";
-import {
-	groupJoinController,
-	sendConfirmShiftFuncController,
-	sendShiftRequestFuncController,
-} from "../controllers/messege.controller";
-import { attachStoreId } from "../middlewares/request/attachStoreId";
-import { attachUserId } from "../middlewares/request/attachUserId";
-const router = express.Router();
-
-router.post("/", groupJoinController);
-router.post(
-	"/request-shift",
-	attachUserId,
-	attachStoreId,
-	sendShiftRequestFuncController,
-);
-router.post(
-	"/confirm-shift",
-	attachUserId,
-	attachStoreId,
-	sendConfirmShiftFuncController,
-);
-
-export default router;
