--- src/validations/message.validation.ts
@@ -1,6 +1,0 @@
-import { z } from "zod";
-
-export const pushMessageValidate = z.object({
-	groupId: z.string().max(250),
-});
-export type pushMessageInput = z.infer<typeof pushMessageValidate>;
